Example trace generator, trace replay, and automatic latency-bound throughput measurement.
To be clear, this code is very rough. I am looking for a high-level review to answer the questions like:
1. Is the code sufficiently general to cover all use cases?
2. What should the interface to this code look like?

I have a bunch of TODOs that might become TODONTs depending on reviewer
feedback. If internal reviewers are relatively happy with this as a first draft,
I would like to get broader feedback from the MLPerf community.

More fine-grained reviews are, of course, also appreciated and very welcome!