LoadGen: Update README and update python build reqs.
Remove the old loadgen README.md and repace it with
README_BUILD.md along with some updates. The old
readme doesn't pertain to the current code very well
and includes a lot of language that will be captured
by the MLPerf inference specification.

pybind11 needs c++14 features, the way loadgen uses it.
Document it as a required CFLAG envrionment variable
in case the python used for building the loadgen module
doesn't use >=c++14 by default.


Former-commit-id: 1db1b6c418b7db365f17304fb0edef50c539a7c9