LoadGen: Change C++ API to pass vector refs
The python API doesn't change, but changing the C++ API
to pass the underlying vectors, rather than pointers to
the data, lets the python bindings code path avoid copying
the data twice.

Also adds a stub for passing latency results to the SUT
if the SUT wants to do inspection and analysis of its
own.