LoadGen: Add lots of knobs to TestSettings.
Adds a bunch of fields to TestSettings to control per-scenario
knobs and override settings for testing purposes.

Adds a TestSettingsInternal struct that handles the spec defaults
and user-defined overrides, sanitizing the settings in a way
that is easy for the loadgen to use.

Renames SearchForQps to FindPeakPerformance since the latter is a
more generic term that also applies to the multi-stream scenario.


Former-commit-id: 308a41780d83678c8b35c60903902df573134d26