LoadGen: Add TestScenario::MultiStreamFree
This is an alternative to MultiStream that allows for
a free running query interval based on completion of
the N'th oldest query.

This is not an official scenario, but is implemented for
evaluation purposes. Compared to MultiStream, it 1) enables
and potentially rewards pipelining and 2) does not quantize
the performance result to a fixed frequency.

It is split out into a separate scenario to simplify
the code for the vanilla MultiStream scenario.


Former-commit-id: f7b2a7d6caa5e393517e99edf5159c82cc569605