LoadGen: Improve version generator git check.
updates the script to explicitly check whether loadgen is part
of a git repository in the same place that the mlperf inference
repository would have it, not just that it's part of any git
repository.

This would still have some false positives, but if someone
doesn't want their project's git commit subjects included
in those cases, we can update the scripts at that time. This
is unlikely to be an actual issue.


Former-commit-id: 632359058dc25d60bfcf1bb9979dd3f941003d2c