LoadGen: Simplify atexit segfault fix.
This also seems to work without requiring exceptions to
be enabled.

There is a risk that std::unique_ptr<TlsLogger>'s destructor
gets unloaded along with the loadgen module, but it doesn't
seem to be an issue in practice. Either it's optimized away
or it's still valid for some reason.


Former-commit-id: d29e1efd2eafe42d1e8b600bcfbd0fe239e008d1