LoadGen: Change QSL contract for sample loading.
When I added support for contiguous samples in multi-stream mode,
I added "padding" samples to the end of LoadSamplesToRam.
To keep the distribution of samples equal, regardless of the
requested samples per query, I made the padding look as if it
wraps around to the beginning. This had the side-effect of
breaking an old contract of LoadSamplesToRam: "The load generator
will never load a currently loaded sample.". In turn, this broke
the unload callback implementation which is now asked to unload
the same sample twice. I had that contract in place originally
in anticipation of potentially streaming samples from disk,
which we later decided against doing.

This contract is no longer important to keep for v0.5.