LoadGen: Add null SUT + QSL perftests in C++.
Results for single stream on my machine:
QPS w/ loadgen overhead  : 412752.87
QPS w/o loadgen overhead : 11111111.11

That's 2.423 microseconds per query with loadgen overhead.

All logging and tracing was enabled. The test ran for
~5 seconds and resulted in a ~1GB trace file.

This was on a desktop with an SSD and the following CPU:
Intel(R) Xeon(R) CPU E5-2690 v3 @ 2.60GHz


Former-commit-id: 31fab7000a9461576089c73c3074421442f7eb32