LoadGen: Sanity checks for latency recording.
Flag errors if samples are missing their latency
or if an attempt to complete a sample twice is made.

Also, take into account the first expected sample_sequence_id
when indexing into the AsyncLog::latencies_ vector.
The sample_sequence_id used to reset at the start of a series,
but it no longer does.


Former-commit-id: c14025543fd0759a2ecbfccdbd9caaa3c0c53635