LoadGen: Simplify logging and tracing interface. (#263)
* LoadGen: Simplify logging interface.

Prior to this patch, logging errors and detail logs was
error prone because we'd have to call matching functions
from the producing thread and on the consumer thread
inside the lambda. Not calling the matching functions
would result in silent unexpected behavior.

This patch:
* Adds LogSummary and LogDetail functions that automatically
  call the corresponding method on the
  AsyncLog using a functor proxy.
* The LogDetail proxy has an Error() method to report errors
  in the detail log.
* The functor proxies also provide access to the underyling
  AsyncLog as an escape hatch.

Using the AsyncLog directly is still supported since it's
useful to log multiple types of messages within a single
lambda.

Note: ScopedTraces have not been coverted yet, but could
benefit from a similar change.

* LoadGen: Simplify tracing interface.

Adds an AsyncTrace proxy that prevents the caller
from needing to know which AsyncLog method to call.

Prior to this patch, calling the wrong AsyncLog method
could silently result in unexpected behavior.