LoadGen: Add tests for "remainder" samples.
Adds a test to verify we don't hang in the
offline scenario in accuracy mode when the
last loadable set is smaller than the rest.

Adds a test to verify MultiStream scenario
has contiguous samples in the accuracy mode
even when the loadable set size is not an
exact multiple of samples_per_query.


Former-commit-id: 74a5eda0a4856bfbc213ee1c7f7d504a525f8b46