Add rows for the MobileNet TFLite models.
Accuracy provided as measured via CK with OpenCV preprocessing:
https://github.com/mlperf/inference/pull/312

The Top 1 accuracy of the non-quantized TFLite model is the same as that
of the non-quantized TF model (71.516%). Potentially it's the same when
measured via the official code (71.68%).

The Top 1 accuracy of the quantized TFLite model, however, is higher
than that of the quantized TF model (70.776% vs. 70.654%, or 61 more
correctly classified images). Interestingly, the Top 5 accuracy is lower
(89.198% vs. 89.514%, or 158 fewer correctly classified images).

3 digits after the floating point are provided as recommended here:
https://github.com/mlperf/inference/issues/293

Partially addresses:
https://github.com/mlperf/inference/issues/289