LoadGen: Add option to log responses in performance mode (#359)
https://github.com/mlperf/inference_policies/issues/41

Description and reasoning behind the approach taken:
To address Brian's concerns, we no longer generate random numbers in SampleComplete(),
since the rng is not thread-safe. Instead of passing a flag through SampleMetadata as per
Brian's proposal, we pass a random number that is now generated in GenerateQueries().
Then in SampleComplete(), we offset the random number by another random number in
ResponseDelegate and check against target interval based on acc_log_probability.
This prevents a SUT from simply reading SampleMetadata to determine which samples
are being logged.