[GNMT] Added BPE file and preprocessing script
In order to translate other text files, preprocessing needs to be done.
This commit contains a BPE file, as well as a preprocessing script that
downloads and runs a tokenizer and subword splitter.

Please note, that subword splitting is data dependent. Therefore, a
better place for the bpe file would be in the same tarball as the one
that donwload_dataset.sh downloads.


Former-commit-id: f6df8eea3238bdc15166ae86eda77d3e96b35a00