loadgen: Make server test duration depend on scheduled time.
Prior to this patch, it depended on the issue time which doesn't
match how the QPS metric is calcualted.

* Genererates exactly the number of queries to issue.
* Attempt to issue all queries generated.
* Update MinDurationMet to depend on scheduled time.

This patch also fixes a bug where a "ran out of queries"
error would spuriously be reported in the corner case
that all queries were used but the minimum durations
happened to be met.