Example ONNX export test
How to run:

python -u $(which pytest) -k test_export_encoder test_onnx.py

Notably, "-u" will make sure that C++ and python print statemetns to
stdout will be in-sync.

There's a bug when we pass a tuple to AudioPreprocessing::forward
where the length Tensor gets the wrong shape (it should be a scalar
when batch size is 1, but intead it gets the same shape as
input_signal), so I change the signature.

This test is currently failing.